E-Procurement project for KARAM Industries
e-Procurement or electronic procurement refers to the process of purchase and sale of goods or services through electronic methods, primarily the Internet. It is an alternative to the manual process of procurement, and is certainly superior to the latter in many respects. Organizations are increasingly opting for e-Procurement platforms, realizing its potential to curb irregularities and unnecessary costs.